# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servidor API para el Proyecto Sakoo (Arquitectura Limpia).",
        "title": "Sakoo Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte Sakoo",
            "url": "http://www.sakoo.com",
            "email": "soporte@sakoo.com"
        },
        "license": {
            "name": "Propietaria",
            "url": "http://www.sakoo.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/payments/commitments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite registrar un nuevo compromiso de pago (POST) o listar todos los compromisos segmentados por estado (por_vencer, vencidos, cumplidos) del usuario (GET).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Compromisos de Pago"
                ],
                "summary": "Gestionar compromisos de pago (Crear / Listar Segmentados)",
                "parameters": [
                    {
                        "description": "Datos del compromiso (requerido solo para POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CommitmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error al procesar la solicitud",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite registrar un nuevo compromiso de pago (POST) o listar todos los compromisos segmentados por estado (por_vencer, vencidos, cumplidos) del usuario (GET).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Compromisos de Pago"
                ],
                "summary": "Gestionar compromisos de pago (Crear / Listar Segmentados)",
                "parameters": [
                    {
                        "description": "Datos del compromiso (requerido solo para POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CommitmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Error al procesar la solicitud",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/commitments/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite actualizar por completo un compromiso de pago (PUT) o eliminarlo de forma lógica (DELETE) especificando su ID en la ruta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Compromisos de Pago"
                ],
                "summary": "Gestionar detalle de compromiso de pago (Actualizar / Eliminar)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID del compromiso de pago",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados del compromiso (requerido solo para PUT)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CommitmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID inválido, no autorizado o error de negocio",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Permite actualizar por completo un compromiso de pago (PUT) o eliminarlo de forma lógica (DELETE) especificando su ID en la ruta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Compromisos de Pago"
                ],
                "summary": "Gestionar detalle de compromiso de pago (Actualizar / Eliminar)",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID del compromiso de pago",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados del compromiso (requerido solo para PUT)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CommitmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID inválido, no autorizado o error de negocio",
                        "schema": {
                            "$ref": "#/definitions/github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_aaron_sakoo-backend_internal_api_response.APIResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "track_code": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.CommitmentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency_id": {
                    "type": "integer"
                },
                "due_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Coloque el token JWT de la siguiente manera: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}